Use hcsshim constants for Windows build numbers
This replaces the hard-coded build numbers with constants that
are now defined in hcsshim.

Note that build 16236 is not defined in the package, and may
be referring to RS2 (which was a client-only release).

Checks for this version were added in commits;

a9db5c53867a33c47c13a4eac7c1a28cc3c35054 (#1900), and
6681c0243cc03669a677efe4dde70976b1bf0746 (#2070)

In a9db5c53867a33c47c13a4eac7c1a28cc3c35054, it is described
that the intent was to allow these options only in Windows RS3
so I changed the comparisons to match RS3 (and up), or disable
the option on anything older than RS3.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>